Rename _dicom_exceptions.py - Rename to public module
